+++ internal/domain/entity/order.go
@@ -1,13 +1,13 @@
 package entity
 
 type Order struct {
-	ID          uint64 `gorm:"primary_key:auto_increment" json:"id" gorm:"column:id"`
-	UserID      uint64 `gorm:"type:varchar(255)" json:"user_id" gorm:"column:user_id"`
+	ID          int    `gorm:"primary_key:auto_increment" json:"id" gorm:"column:id"`
+	UserID      int    `gorm:"type:varchar(255)" json:"user_id" gorm:"column:user_id"`
 	Description string `gorm:"type:varchar(255)" json:"description" gorm:"column:description"`
 }
 
 type OrderPhoto struct {
-	ID      uint64 `gorm:"primary_key:auto_increment" json:"id" gorm:"column:id"`
-	OrderID uint64 `gorm:"type:varchar(255)" json:"order_id" gorm:"column:order_id"`
-	PhotoID uint64 `gorm:"type:varchar(255)" json:"photo_id" gorm:"column:photo_id"`
+	ID      int `gorm:"primary_key:auto_increment" json:"id" gorm:"column:id"`
+	OrderID int `gorm:"type:varchar(255)" json:"order_id" gorm:"column:order_id"`
+	PhotoID int `gorm:"type:varchar(255)" json:"photo_id" gorm:"column:photo_id"`
 }
